Added code for key validation
The key which the user inputs can be passed on to this script using GET or POST method (GET is used by default for testing purposes). This script will check the input key with the hash values stored in the database and output '1' if the key is valid, '0' if the key is invalid.